requirements.txt
# 基礎套件
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.1
numpy==1.24.3

# 數據處理
python-dotenv==1.0.0
schedule==1.2.0

# API 相關
alpha-vantage==2.3.1

# 網頁爬蟲
undetected-chromedriver==3.5.4
lxml==4.9.3

# 日誌和配置
loguru==0.7.2

# 測試
pytest==7.4.3

# 配置管理
python-dotenv==1.0.0

# HTTP客戶端
httpx==0.25.0

# 時間處理
python-dateutil==2.8.2

# 數據驗證
pydantic==2.4.2

# 重試機制
tenacity==8.2.3

# 隨機延遲
fake-useragent==1.4.0

# Web框架
Flask==3.0.0
Werkzeug==3.0.1

# JSON處理
orjson==3.9.10 